+++ spec/lib/facebook_ads/ad_objects/server_side/http_util_spec.rb
@@ -1,0 +1,37 @@
+# Copyright (c) 2017-present, Facebook, Inc. All rights reserved.
+#
+# You are hereby granted a non-exclusive, worldwide, royalty-free license to use,
+# copy, modify, and distribute this software in source code or binary form for use
+# in connection with the web services and APIs provided by Facebook.
+#
+# As with any software that integrates with the Facebook platform, your use of
+# this software is subject to the Facebook Platform Policy
+# [http://developers.facebook.com/policy/]. This copyright notice shall be
+# included in all copies or substantial portions of the software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+require 'openssl'
+
+RSpec.describe 'FacebookAds::ServerSide::HttpUtil' do
+    context 'app_secret_proof' do
+        it 'generates the proof' do
+            app_secret = 'app-secret-1'
+            access_token = 'access-token-2'
+            digest = OpenSSL::Digest.new('sha256')
+            expected = OpenSSL::HMAC.hexdigest(
+                digest,
+                app_secret,
+                access_token
+            )
+
+            actual = FacebookAds::ServerSide::HttpUtil.appsecret_proof(app_secret, access_token)
+            expect(actual).to eq(expected)
+        end
+    end
+end