+++ lib/facebook_ads/ad_objects/product_item.rb
@@ -344,6 +344,7 @@ module FacebookAds
     field :url, 'string'
     field :visibility, { enum: -> { VISIBILITY }}
     field :additional_image_files, { list: 'file' }
+    field :additional_uploaded_image_ids, { list: 'string' }
     field :android_app_name, 'string'
     field :android_class, 'string'
     field :android_package, 'string'